+++ src/AppInstallerCLITests/Dependencies.cpp
@@ -9,7 +9,6 @@
 #include <AppInstallerStrings.h>
 #include <Workflows/DependenciesFlow.h>
 #include <Workflows/WorkflowBase.h>
-#include <winget/Filesystem.h>
 #include <winget/ManifestYamlParser.h>
 #include <winget/PathVariable.h>
 #include <winget/RepositorySource.h>
@@ -209,101 +208,4 @@ TEST_CASE("DependencyNodeProcessor_NoMatches", "[dependencies]")
     REQUIRE(dependencyList.Size() == 0);
     REQUIRE(installOutput.str().find(Resource::LocString(Resource::String::DependenciesFlowNoMatches)) != std::string::npos);
     REQUIRE(result == DependencyNodeProcessorResult::Error);
-}
-
-std::wstring GetCurrentProcessPathVariable()
-{
-    size_t requiredSize;
-    _wgetenv_s(&requiredSize, nullptr, 0, L"PATH");
-
-    if (requiredSize > 0)
-    {
-        auto buffer = std::make_unique<wchar_t[]>(requiredSize);
-        errno_t errorResult = _wgetenv_s(&requiredSize, buffer.get(), requiredSize, L"PATH");
-        if (errorResult == 0)
-        {
-            return std::wstring(buffer.get());
-        }
-    }
-    return {};
-}
-
-TEST_CASE("RefreshEnvironmentVariable_User", "[dependencies]")
-{
-    if (!AppInstaller::Runtime::IsRunningAsAdmin())
-    {
-        WARN("Test requires admin privilege. Skipped.");
-        return;
-    }
-
-    std::wstring testPathEntry = L"testUserPathEntry";
-    auto pathVariable = AppInstaller::Registry::Environment::PathVariable(ScopeEnum::User);
-    pathVariable.Append(testPathEntry);
-
-    std::wstring initialPathValue = GetCurrentProcessPathVariable();
-    bool firstCheck = initialPathValue.find(testPathEntry) != std::string::npos;
-
-    AppInstaller::Registry::Environment::RefreshPathVariableForCurrentProcess();
-
-    std::wstring updatedPathValue = GetCurrentProcessPathVariable();
-    bool secondCheck = updatedPathValue.find(testPathEntry) != std::string::npos;
-
-    pathVariable.Remove(testPathEntry);
-
-    REQUIRE_FALSE(firstCheck);
-    REQUIRE(secondCheck);
-}
-
-TEST_CASE("RefreshEnvironmentVariable_System", "[dependencies]")
-{
-    if (!AppInstaller::Runtime::IsRunningAsAdmin())
-    {
-        WARN("Test requires admin privilege. Skipped.");
-        return;
-    }
-
-    std::wstring testPathEntry = L"testSystemPathEntry";
-    auto pathVariable = AppInstaller::Registry::Environment::PathVariable(ScopeEnum::Machine);
-    pathVariable.Append(testPathEntry);
-
-    std::wstring initialPathValue = GetCurrentProcessPathVariable();
-    bool firstCheck = initialPathValue.find(testPathEntry) != std::string::npos;
-
-    AppInstaller::Registry::Environment::RefreshPathVariableForCurrentProcess();
-
-    std::wstring updatedPathValue = GetCurrentProcessPathVariable();
-    bool secondCheck = updatedPathValue.find(testPathEntry) != std::string::npos;
-
-    pathVariable.Remove(testPathEntry);
-
-    REQUIRE_FALSE(firstCheck);
-    REQUIRE(secondCheck);
-}
-
-TEST_CASE("VerifyPathRefreshExpandsValues", "[dependencies]")
-{
-    if (!AppInstaller::Runtime::IsRunningAsAdmin())
-    {
-        WARN("Test requires admin privilege. Skipped.");
-        return;
-    }
-
-    std::filesystem::path testEntry{ "%USERPROFILE%\\testPath" };
-    auto pathVariable = AppInstaller::Registry::Environment::PathVariable(ScopeEnum::User);
-    pathVariable.Append(testEntry);
-
-    std::wstring initialPathValue = GetCurrentProcessPathVariable();
-    bool firstCheck = initialPathValue.find(testEntry) != std::string::npos;
-
-    AppInstaller::Registry::Environment::RefreshPathVariableForCurrentProcess();
-
-    // %USERPROFILE% should be replaced with the actual path.
-    std::wstring updatedPathValue = GetCurrentProcessPathVariable();
-    std::wstring expandedTestPath = AppInstaller::Filesystem::GetExpandedPath(testEntry.u8string());
-    bool secondCheck = updatedPathValue.find(expandedTestPath) != std::string::npos;
-
-    pathVariable.Remove(testEntry);
-
-    REQUIRE_FALSE(firstCheck);
-    REQUIRE(secondCheck);
 }
